preview-mode: P: Preview file at cursor in non-focused preview window
Following up on the PR from @blueyed (with additional minor changes),
I personally believe that this is how a preview functionality should
behave i.e. file under cursor opens in a non-focused split preview
window with properties of a scratch buffer in order to avoid
unintentional file modifications and buffer list clutter.

The current preview-mode or preview-workflow implemented by @justinmk,
particularly the mapping `p` (lowercase) in normal mode, seems to be
redundant as its behaviour is already fulfilled by the mapping `a` in
normal mode to explicitly open file in a new vertical window. Thus, it
may be more sensible to not have previewed file under cursor to be
focused-split, listed and modifiable.

See also: #65, #68